[refactor] (models.py migrations/) moved two classes from migration app to testapp2
